+++ pkg/helper/seed.go
@@ -2,7 +2,6 @@ package helper
 
 import (
 	"github.com/gardener/gardener/pkg/apis/core/v1beta1"
-	operatorv1alpha1 "github.com/gardener/gardener/pkg/apis/operator/v1alpha1"
 )
 
 // GetSeedSpecificAllowedCIDRs returns the node and pod CIDRs from the seed
@@ -24,12 +23,4 @@ func GetSeedIngressDomain(seed *v1beta1.Seed) string {
 		domain = seed.Spec.Ingress.Domain
 	}
 	return domain
-}
-
-// GetGardenSpecificAllowedCIDRs returns the node and pod CIDRs from the garden runtime cluster
-func GetGardenSpecificAllowedCIDRs(garden *operatorv1alpha1.Garden) []string {
-	cidrs := make([]string, 0)
-	cidrs = append(cidrs, garden.Spec.RuntimeCluster.Networking.Nodes...)
-	cidrs = append(cidrs, garden.Spec.RuntimeCluster.Networking.Pods...)
-	return cidrs
 }
